requirements.txt
affine==2.3.1
anyio==3.6.1
asgiref==3.5.2
astroid==2.9.3
async-generator==1.10
attrs==21.4.0
beautifulsoup4==4.10.0
branca==0.5.0
CacheControl==0.12.10
cachetools==5.2.0
cachy==0.3.0
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
cleo==0.8.1
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
clikit==0.6.2
collections-extended==2.0.2
colorama==0.4.4
crashtest==0.3.1
distlib==0.3.4
earthengine-api==0.1.312
fake-useragent==0.1.11
fastapi==0.78.0
filelock==3.6.0
folium==0.12.1.post1
future==0.18.2
h11==0.13.0
h2==4.1.0
hpack==4.0.0
html5lib==1.1
httplib2==0.20.4
httplib2shim==0.0.3
httptools==0.4.0
#hyper==0.7.0
hypercorn==0.13.2
hyperframe==6.0.1
idna>=2.5
importlib-metadata==4.11.3
isort==5.10.1
Jinja2==3.1.2
keyring==23.5.0
lazy-object-proxy==1.7.1
lockfile==0.12.2
MarkupSafe==2.1.1
mccabe==0.6.1
msgpack==1.0.3
numpy==1.22.4
opencv-python==4.6.0.66
outcome==1.2.0
packaging==20.9
pastel==0.2.1
pexpect==4.8.0
pip==22.1.2
pkginfo==1.8.2
platformdirs==2.5.1
poetry-core==1.0.8
priority==2.0.0
protobuf==3.20.1
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.9.1
pylev==1.4.0
pylint==2.12.2
pyparsing==3.0.7
python-dotenv==0.19.2
pywin32-ctypes==0.2.0
PyYAML==6.0
requests==2.28.0
requests-toolbelt==0.9.1
rsa==4.8
sentinelsat==1.1.1
setuptools==62.3.4
shellingham==1.4.0
six==1.16.0
sniffio==1.2.0
snuggs==1.4.7
sortedcontainers==2.4.0
soupsieve==2.3.1
starlette==0.19.1
toml==0.10.2
tomlkit==0.10.0
tqdm==4.64.0
trio==0.21.0
typing_extensions==4.2.0
uritemplate==3.0.1
urllib3==1.26.9
uvicorn==0.17.6
virtualenv==20.13.4
watchgod==0.8.2
webencodings==0.5.1
websockets==10.3
wrapt==1.13.3
wsproto==1.1.0
zipp==3.7.0
wheel
SQLAlchemy
alembic
"python-jose[cryptography]"
"passlib[bcrypt]"
python-multipart